requirements.txt
cffi==1.13.1
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.6
django-cryptographic-fields==0.5.1
djangorestframework==3.10.3
gunicorn==19.9.0
pycparser==2.19
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
